Linked buttons to pages and rearranged folder structure
