+++ app/index.tsx
@@ -2,5 +2,5 @@ import { Redirect } from 'expo-router';
 import 'react-native-url-polyfill/auto';
 
 export default function Index() {
-  return <Redirect href="/login" />;
+  return <Redirect href="/screens/login" />;
 }
